Testes finais: status: projeto funcionando
