--- mudi/src/main/java/br/com/mudi/LoginController.java
@@ -1,16 +1,0 @@
-package br.com.mudi.controllers;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-public class LoginController {
-
-    @GetMapping
-    @RequestMapping("/login")
-    public String login() {
-        return "login";
-    }
-    
-}
