+++ mudi/src/main/java/br/com/mudi/MudiApplication.java
@@ -1,5 +1,4 @@
 package br.com.mudi;
-
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
@@ -8,6 +7,6 @@ public class MudiApplication {
 
 	public static void main(String[] args) {
 		SpringApplication.run(MudiApplication.class, args);
-	}
 
+	}
 }
