+++ mudi/src/main/java/br/com/mudi/controllers/LoginController.java
@@ -1,4 +1,14 @@
 package br.com.mudi.controllers;
 
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
 public class LoginController {
+    @GetMapping
+    @RequestMapping("/login")
+    public String login() {
+        return "login";
+    }
 }
