+++ mudi/src/main/java/br/com/mudi/controllers/RequestController.java
@@ -21,4 +21,13 @@ public class RequestController {
         return "formCad";
     }
 
+    @PostMapping
+    public String insert(@Valid Request request, BindingResult result){
+        if(result.hasErrors()) {
+            return "formCad";
+        }
+        request.setStatusRequest(StatusRequest.valueOf("WAITING"));
+        rr.save(request);
+        return "redirect:/home";
+    }
 }
